refactor so django does not import serverless code
