+++ documentcloud/common/serverless/utils.py
@@ -5,14 +5,12 @@ Helper functions to perform common serverless operations.
 # Standard Library
 import json
 import logging
-import math
 import time
 from urllib.parse import urljoin
 
 # Third Party
 import environ
 import furl
-import pymupdf
 import redis as _redis
 import requests
 from redis.lock import Lock
@@ -350,27 +348,3 @@ def initialize_text_positions(redis, doc_id, page_count):
     pipeline.setbit(text_position_bit_field, page_count - 1, 0)
     pipeline.expire(text_position_bit_field, REDIS_TTL)
     pipeline.execute()
-
-
-def graft_page(positions, pdf_page):
-    """Graft words with position information onto a PDF page"""
-
-    default_fontsize = 15
-
-    for position in positions:
-        word_text = position["text"]
-        text_length = pymupdf.get_text_length(
-            word_text,
-            fontsize=default_fontsize,
-        )
-        width = (position["x2"] - position["x1"]) * pdf_page.rect.width
-        fontsize_optimal = int(math.floor((width / text_length) * default_fontsize))
-        pdf_page.insert_text(
-            point=pymupdf.Point(
-                position["x1"] * pdf_page.rect.width,
-                position["y2"] * pdf_page.rect.height,
-            ),
-            text=word_text,
-            fontsize=fontsize_optimal,
-            fill_opacity=0,
-        )
